Reinspection Protection
Added an ArrayList of hashcodes for objects inspected with Inspector. When Inspect is called, it checks if the hashcode is contained in the list of visited objects, skipping inspection is the list contains it. If not, it adds the objects hashcode to the list and continues inspection. This is useful if there are two objects that reference each other, without a guard such as this, they would loop infinitely in recursive inspection.