Unit test bugfix
testParameter1 had a bug where tests would fail due to getDeclaredMethods returning them in an order that was not predicted. To fix this, all parameter strings from method parameters are stored in an ArrayList, which is checked after to ensure the parameter list for each method is present and properly formatted.